requirements.txt
streamlit
groq
python-dotenv
pandas
requests